+++ viscy/translation/evaluation.py
@@ -39,7 +39,10 @@ class SegmentationMetrics2D(LightningModule):
                 "accuracy": (accuracy(pred_binary, target_binary, task="binary")),
                 "dice": (
                     dice_score(
-                        pred_binary, target_binary, num_classes=2, input_format="index"
+                        pred_binary.long(),
+                        target_binary.long(),
+                        num_classes=2,
+                        input_format="index",
                     )
                 ),
                 "jaccard": (jaccard_index(pred_binary, target_binary, task="binary")),
